+++ src/main/java/jpashop/jpashop/service/OrderService.java
@@ -8,8 +8,8 @@ import jpashop.jpashop.domain.Member;
 import jpashop.jpashop.domain.Order;
 import jpashop.jpashop.domain.OrderItem;
 import jpashop.jpashop.dto.order.OrderDTO;
-import jpashop.jpashop.dto.order.form.OrderAddDTO;
-import jpashop.jpashop.dto.order.form.OrderAddDTOList;
+import jpashop.jpashop.dto.order.OrderAddDTO;
+import jpashop.jpashop.dto.order.OrderAddDTOList;
 import jpashop.jpashop.repository.ItemRepository;
 import jpashop.jpashop.repository.MemberRepository;
 import jpashop.jpashop.repository.OrderItemRepository;
